* changed the check scripts to accept options -p and -r as documented in BNC#893131#c2   This was done in a way that the output of the scripts will be the same   when compared to having called the older versions of the scripts with just $MD5.   The scripts can not be called by just $MD5 anymore (meaning, we need to change the caller in mtui accordingly). * rewrote doit.sh for BNC#89313 and to pair the checker script with the compare script * updated the README

+++ helper/check_dependencies.sh
@@ -1,9 +1,38 @@
 #!/bin/bash
 
+export LANG=C
+
+help () {
+   cat <<EOF
+
+usage: $0 <id>
+
+where id is either \$md5sum or (openSUSE|SUSE):Maintenance:\$issue:\$request
+
+$0 checks the rpm dependencies of all installed packages, but with --nofiles
+
+EOF
+}
+
+while getopts ":r:p:h" opt
+do
+   case $opt in
+   h) help="true" ;;
+   \?) echo "INFO: ignoring unsupported option $OPTARG" ;;
+   esac
+done
+
+id=$BASH_ARGV
+
+if [ -n "$help" ]; then
+   help
+   exit 0
+fi
+
 (while true; do sleep 60; echo "."; done) &
 PONG=$!
 
-rpm -Va --nofiles | sort
+rpm -Va --nofiles | sort >&2
 
 kill $PONG &>/dev/null
